requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.1.1
email-validator==2.3.0
requests==2.32.5
Werkzeug==3.1.3
bcrypt==4.2.1